requirements.txt
# langchain==0.1.1
# langchain-community==0.0.13
# langchain-core==0.1.12
# langsmith==0.0.83
uvicorn[standard]
fastapi
python-dotenv
# torch==2.1.2
# vllm==0.2.7